requirements.txt
Flask-Bootstrap
mysql-connector
numpy
